Crate gloo_storage

This crate provides wrappers for the Web Storage API

The data is stored in JSON form. We use serde for serialization and deserialization.

Modules

errors

All the errors.

Structs

LocalStorage

Provides API to deal with localStorage

SessionStorage

Provides API to deal with sessionStorage

Traits

Storage

Trait which provides implementations for managing storage in the browser.

Type Definitions

Result

gloo-storage’s Result